Fix alternating taplistio sends (Fixes #201)
